LPS-57049 Move classes
