+++ modules/apps/export-import/export-import-service/src/com/liferay/exportimport/lar/CurrentUserIdStrategy.java
@@ -12,11 +12,12 @@
  * details.
  */
 
-package com.liferay.portlet.exportimport.lar;
+package com.liferay.exportimport.lar;
 
 import com.liferay.portal.kernel.util.Validator;
 import com.liferay.portal.model.User;
 import com.liferay.portal.service.UserLocalServiceUtil;
+import com.liferay.portlet.exportimport.lar.UserIdStrategy;
 
 /**
  * @author Bruno Farache
