+++ modules/apps/export-import/export-import-service/src/com/liferay/exportimport/lar/PermissionImporter.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portlet.exportimport.lar;
+package com.liferay.exportimport.lar;
 
 import com.liferay.portal.NoSuchTeamException;
 import com.liferay.portal.kernel.util.GetterUtil;
@@ -34,6 +34,8 @@ import com.liferay.portal.service.RoleLocalServiceUtil;
 import com.liferay.portal.service.ServiceContext;
 import com.liferay.portal.service.TeamLocalServiceUtil;
 import com.liferay.portal.service.permission.PortletPermissionUtil;
+import com.liferay.portlet.exportimport.lar.ExportImportPathUtil;
+import com.liferay.portlet.exportimport.lar.PortletDataContext;
 
 import java.util.ArrayList;
 import java.util.HashMap;
